server/requirements.txt
Flask>=3.0.0,<4.0.0
Flask-Cors>=4.0.0,<5.0.0
pymongo>=4.6.0,<5.0.0
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pytest>=7.4.0,<8.0.0

